% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altered_dunn.test.R
\name{Adunn.test}
\alias{Adunn.test}
\title{Dunns multicomparison test from dunn.test library}
\usage{
Adunn.test(
  x = NA,
  g = NA,
  method = p.adjustment.methods,
  kw = TRUE,
  label = TRUE,
  wrap = FALSE,
  table = TRUE,
  list = FALSE,
  rmc = FALSE,
  alpha = 0.05,
  altp = FALSE
)
}
\description{
Dunns multicomparison test from dunn.test library
}
