% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normtest.R, R/normtest_2.R
\name{basic_stats}
\alias{basic_stats}
\title{Basic Statistics}
\usage{
basic_stats(x, ...)

basic_stats(x, ...)
}
\arguments{
\item{x}{Numeric Vector}

\item{...}{Any parameters added to basic_stats}

\item{x}{Numeric Vector}

\item{...}{Any parameters added to basic_stats}
}
\value{
Numeric vector of length six: mean, sd, skewness, kurtosis, n, n_na

Numeric vector of length six: mean, sd, skewness, kurtosis, n, n_na
}
\description{
Basic Statistics

Basic Statistics
}
\examples{
set.seed(666)
x <- rnorm(10, 7, 1)
basic_stats(x)
set.seed(666)
x <- rnorm(10, 7, 1)
basic_stats(x)
}
