% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds.R
\name{check_means}
\alias{check_means}
\title{check if two or more group means are zero}
\usage{
check_means(my_data, var1)
}
\arguments{
\item{my_data}{a data object}

\item{var1}{a list that points column names of my_data to the factors of the experiment}
}
\description{
check if two or more group means are zero
}
