% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds.R
\name{chop}
\alias{chop}
\title{Chop multi group result files into single comparisons.}
\usage{
chop(x, p = 0.05, f = 1)
}
\arguments{
\item{x}{a data frame, usually one produced by running the main multigroup normtest function, containing only groupwise comparisons.}

\item{p}{p value cutoff for significance. Defaults to 0.05}

\item{f}{log2 fold change cut off. Defaults to 1}
}
\description{
Chop multi group result files into single comparisons.
}
