% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fold_change.R
\name{combine_fc_pvals}
\alias{combine_fc_pvals}
\title{This function is a helper function for the statomatic workflow. It is not really designed to be used by itself. Combines columns of p value and fold change data.frames}
\usage{
combine_fc_pvals(a, b)
}
\arguments{
\item{a}{a data.frame, these values will be on the left}

\item{b}{a data.frame, these values will be on the right}
}
\description{
This function is a helper function for the statomatic workflow. It is not really designed to be used by itself. Combines columns of p value and fold change data.frames
}
