% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fold_change.R
\name{fold_change}
\alias{fold_change}
\title{Calculates fold changes, defaults to log2, for each row of a data frame according to specified factors}
\usage{
fold_change(my_data, var1, FUN = log2, ...)
}
\arguments{
\item{my_data}{a numeric data.frame}

\item{var1}{a list which maps columns of my_data to experimental factors
FUN a function which can be passed one numeric value and return one numeric value. Default is log2, but could be log, identity, function(x) {return(x * 2)}, etc.}
}
\value{
a data.frame of fold changes
}
\description{
Calculates fold changes, defaults to log2, for each row of a data frame according to specified factors
}
