% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normtest.R, R/normtest_2.R
\name{ks}
\alias{ks}
\title{Title Kolmogorov Smirnov Test}
\usage{
ks(x, ...)

ks(x, ...)
}
\arguments{
\item{x}{Numeric Vector}

\item{...}{Any extra param}

\item{x}{Numeric Vector}

\item{...}{Any extra param}
}
\value{
list

list
}
\description{
Title Kolmogorov Smirnov Test

Title Kolmogorov Smirnov Test
}
\examples{
set.seed(666)
ks_test <- ks(rnorm(10))
x <- rnorm(10, 7, 1)
ks(x)
set.seed(666)
ks_test <- ks(rnorm(10))
x <- rnorm(10, 7, 1)
ks(x)
}
