% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_qqplot.R
\name{make_qqplot}
\alias{make_qqplot}
\title{make_qqplot}
\usage{
make_qqplot(data)
}
\arguments{
\item{data, }{DATAFRAME where data for each continous variable is in its respective column}
}
\value{
LIST of plots where the name of each item are the column name (or index)
}
\description{
Create QQ-plot for each continuous variable in the data
}
\examples{
iris_data <- data.frame("length" = c(1,2,3,4), "width" = c(5,6,7,8))
make_qqplot(iris_data)
}
