% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_setup.R
\name{make_sds}
\alias{make_sds}
\title{Create a StatomaticDataset (sds)}
\usage{
make_sds(x, colData = NULL, rowData = NULL, design = ~1)
}
\arguments{
\item{x}{a numeric matrix or data.frame}

\item{colData}{a data.frame of mappings for columns of x to experimental conditions}

\item{rowData}{row metadata}

\item{design}{a design formula containing column names from colData}
}
\value{
a StatomaticDataset
}
\description{
Create a StatomaticDataset (sds)
}
