% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multigroup_main.R
\name{multigroup_main}
\alias{multigroup_main}
\title{This function puts together all the multigroup functions from the package in the correct order.}
\usage{
multigroup_main(my_data, var1, var2, var3, padj = FALSE)
}
\arguments{
\item{my_data}{a data object}

\item{var1}{a list that points column names of my_data to the factors of the experiment}

\item{var2}{a list that points column names of my_data to the factors of the experiment}

\item{var3}{a list that points column names of my_data to the factors of the experiment}
}
\description{
This function puts together all the multigroup functions from the package in the correct order.
}
