% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normtest.R, R/normtest_2.R
\name{normtest}
\alias{normtest}
\title{Main Normtest Function}
\usage{
normtest(x, ...)

normtest(x, ...)
}
\arguments{
\item{x}{numeric vector}

\item{...}{any argument for inner function basic_stats}

\item{x}{numeric vector}

\item{...}{any argument for inner function basic_stats}
}
\value{
list of class normtest

list of class normtest
}
\description{
Main Normtest Function

Main Normtest Function
}
\examples{
# Basic Examples
set.seed(101)
x <- rnorm(100, 7, 1)
plot.normtest(x)
test <- normtest(x)
test
summary(test)

# Examples with na
y <- x; y[7] <- NA
plot.normtest(y)
normtest(y, na.rm = TRUE)

# non normal examples
z <- rlnorm(100, 7, 2)
plot.normtest(z)
normtest(z)
# Basic Examples
set.seed(101)
x <- rnorm(100, 7, 1)
plot.normtest(x)
test <- normtest(x)
test
summary(test)

# Examples with na
y <- x; y[7] <- NA
plot.normtest(y)
normtest(y, na.rm = TRUE)

# non normal examples
z <- rlnorm(100, 7, 2)
plot.normtest(z)
normtest(z)
}
