% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altered_dunn.test.R
\docType{data}
\name{p.adjustment.methods}
\alias{p.adjustment.methods}
\title{Dunns multicomparison test from dunn.test library}
\format{
An object of class \code{character} of length 8.
}
\usage{
p.adjustment.methods
}
\description{
Dunns multicomparison test from dunn.test library
}
\keyword{datasets}
