% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normtest.R, R/normtest_2.R
\name{plot.normtest}
\alias{plot.normtest}
\title{Plot method for objects of class normtest}
\usage{
\method{plot}{normtest}(x)

\method{plot}{normtest}(x)
}
\arguments{
\item{x}{Object of class normtest}

\item{x}{Object of class normtest}
}
\value{
invisible NULL

invisible NULL
}
\description{
Plot method for objects of class normtest

Plot method for objects of class normtest
}
\examples{
a <- rnorm(1000,7,2)
nt <- normtest(a)
plot(nt)
a <- rnorm(1000,7,2)
nt <- normtest(a)
plot(nt)
}
