% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normtest.R, R/normtest_2.R
\name{plot_normtest}
\alias{plot_normtest}
\title{Plot function for normality test}
\usage{
plot_normtest(x)

plot_normtest(x)
}
\arguments{
\item{x}{numeric vector}

\item{x}{numeric vector}
}
\value{
object of class ggplot

object of class ggplot
}
\description{
Plot function for normality test

Plot function for normality test
}
\examples{
plot_normtest(rnorm(20))
plot_normtest(rnorm(20))
}
