% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normtest.R, R/normtest_2.R
\name{print.normtest}
\alias{print.normtest}
\title{Print Method for Objects of Class normtest}
\usage{
\method{print}{normtest}(x)

\method{print}{normtest}(x)
}
\arguments{
\item{x}{Object of Class normtest}

\item{x}{Object of Class normtest}
}
\value{
invisible NULL

invisible NULL
}
\description{
Print Method for Objects of Class normtest

Print Method for Objects of Class normtest
}
