% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rui.R
\name{rui}
\alias{rui}
\title{This function puts together all the function from the package in the correct order.}
\usage{
rui(my_data, var1, var2, var3)
}
\arguments{
\item{my_data}{a data object}

\item{var1}{a list that points column names of my_data to the factors of the experiment}

\item{var2}{a list that points column names of my_data to the factors of the experiment}

\item{var3}{a list that points column names of my_data to the factors of the experiment}
}
\description{
This function puts together all the function from the package in the correct order.
}
