% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_anova.R
\name{run_anova}
\alias{run_anova}
\title{Performs one or two way anova on each row of a data.frame and reports the p value}
\usage{
run_anova(my_data, var1, var2 = NULL, padj = FALSE)
}
\arguments{
\item{my_data}{a numeric data.frame}

\item{var1}{a list which maps columns of my_data to experimental factors}

\item{var2}{same as var1, but defaults to NULL. using var1 and var2 result in two way anova (if both are valid)}

\item{padj}{boolean, whether or not to include Benjamini Hochberg correction. defaults to FALSE}
}
\value{
a data.frame of p values
}
\description{
Performs one or two way anova on each row of a data.frame and reports the p value
}
