% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/welch_dunnett.R
\name{run_dunnett}
\alias{run_dunnett}
\title{Runs DunnettT3 multi comparison test on each row of my_data}
\usage{
run_dunnett(my_data, var1)
}
\arguments{
\item{my_data}{a data object}

\item{var1}{a list that points column names of my_data to the factors of the experiment}
}
\value{
a data object containing p values for each groupwise comparison as specified by var1
}
\description{
Runs DunnettT3 multi comparison test on each row of my_data
}
