% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kruskalwallis.R
\name{run_kruskal}
\alias{run_kruskal}
\title{Performs Kruskal-Wallis test on each row of a data.frame and reports the p value}
\usage{
run_kruskal(my_data, var1, padj = FALSE)
}
\arguments{
\item{my_data}{a numeric data.frame}

\item{var1}{a list which maps columns of my_data to experimental factors}

\item{padj}{boolean, whether or not to include Benjamini Hochberg correction. defaults to FALSE}
}
\value{
a data.frame of p values
}
\description{
Performs Kruskal-Wallis test on each row of a data.frame and reports the p value
}
