% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_test.R
\name{run_ttest}
\alias{run_ttest}
\title{Performs a t-test on each row of a data.frame.}
\usage{
run_ttest(my_data, var1, padj = FALSE)
}
\arguments{
\item{my_data}{a numeric data.frame}

\item{var1}{a list which maps columns of my_data to experimental factors. This function can handle exactly two experimental factos}

\item{padj}{boolean, whether or not to include Benjamini Hochberg correction. defaults to FALSE}
}
\value{
a data.frame of p values
}
\description{
Performs a t-test on each row of a data.frame.
}
