% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/welch_dunnett.R
\name{run_welch}
\alias{run_welch}
\title{Runs a Welch t-test on each row of my_data}
\usage{
run_welch(my_data, var1)
}
\arguments{
\item{my_data}{a data object}

\item{var1}{a list that points column names of my_data to the factors of the experiment}
}
\value{
a data object with a p value for each row of my_data, computed using welch.test
}
\description{
Runs a Welch t-test on each row of my_data
}
