% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sds_analyze.R
\name{sds_analyze}
\alias{sds_analyze}
\title{The "main" statomatic function. Takes an StatomaticDataset (sds) and performs all applicable statomatic analysis}
\usage{
sds_analyze(object, condition1 = NULL, condition2 = NULL, padj = FALSE)
}
\arguments{
\item{object}{An object of class "StatomaticDataset"}

\item{condition1}{a list which maps columns of get_counts(sds) to experimental factors. If not specified, defaults to NULL and will use the design
formula of sds}

\item{condition2}{same as condition1. May be specified for a multi-factor experiment design, but defaults to NULL and if design formula is only of
length 1, this condition is not used.}

\item{padj}{boolean, whether or not to include Benjamini Hochberg correction. defaults to FALSE}
}
\value{
a description of the method
}
\description{
The "main" statomatic function. Takes an StatomaticDataset (sds) and performs all applicable statomatic analysis
}
