% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{sds_anova}
\alias{sds_anova}
\title{Runs anova on data contained in a StatomaticDataset}
\usage{
sds_anova(object, condition1 = NULL, condition2 = NULL, padj = FALSE)
}
\arguments{
\item{object}{A StatomaticDataset}

\item{condition1}{a list which maps columns of get_counts(sds) to experimental factors. If not specified, defaults to NULL and will use the design
formula of sds}

\item{condition2}{same as condition1. if specified here or in the sds design formula will result in a two way anova}

\item{padj}{boolean, whether or not to include Benjamini Hochberg correction. defaults to FALSE}
}
\value{
a description of the method
}
\description{
Runs anova on data contained in a StatomaticDataset
}
