% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{sds_dunntest}
\alias{sds_dunntest}
\title{Runs Dunn test on data contained in a StatomaticDataset}
\usage{
sds_dunntest(object, condition = NULL)
}
\arguments{
\item{object}{a StatomaticDataset}

\item{condition}{a list which maps columns of get_counts(sds) to experimental factors. If not specified, defaults to NULL and will use the design
formula of sds}
}
\value{
a description of the method
}
\description{
Runs Dunn test on data contained in a StatomaticDataset
}
