% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{sds_foldchange}
\alias{sds_foldchange}
\title{Runs fold_change function on data contained in a StatomaticDataset}
\usage{
sds_foldchange(object, condition = NULL)
}
\arguments{
\item{object}{A StatomaticDataset}

\item{condition}{The experimental factor to consider for calculating fold changes. Defaults to NULL and will use the design contained in sds}
}
\value{
A description of the method
}
\description{
Runs fold_change function on data contained in a StatomaticDataset
}
