% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{sds_testnorm}
\alias{sds_testnorm}
\title{Runs the test_norm function on get_counts(sds)}
\usage{
sds_testnorm(object, condition1 = NULL, condition2 = NULL)
}
\arguments{
\item{object}{A StatomaticDataset}

\item{condition1}{a list which maps columns of get_counts(sds) to experimental factors. If not specified, defaults to NULL and will use the design
formula of sds}

\item{condition2}{same as condition1. May be specified for a multi-factor experiment design, but defaults to NULL and if design formula is only of
length 1, this condition is not used.}
}
\value{
a description of the method
}
\description{
Runs the test_norm function on get_counts(sds)
}
