% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{sds_ttest}
\alias{sds_ttest}
\title{Runs t-test on data contained in a StatomaticDataset}
\usage{
sds_ttest(object, condition = NULL, padj = FALSE)
}
\arguments{
\item{object}{a StatomaticDataset}

\item{padj}{boolean, whether or not to include Benjamini Hochberg correction. defaults to FALSE}
}
\value{
a description of the method
}
\description{
Runs t-test on data contained in a StatomaticDataset
}
