% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{sds_tukeytest}
\alias{sds_tukeytest}
\title{Runs tukey-test on data in a StatomaticDataset}
\usage{
sds_tukeytest(object, condition = NULL)
}
\arguments{
\item{object}{a StatomaticDataset}

\item{condition}{a list which maps columns of get_counts(sds) to experimental factors. If not specified, defaults to NULL and will use the design
formula of sds}
}
\value{
a description of the method
}
\description{
Runs tukey-test on data in a StatomaticDataset
}
