% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{sds_wilcoxtest}
\alias{sds_wilcoxtest}
\title{Runs wilcox test on data contained in a StatomaticDataset}
\usage{
sds_wilcoxtest(object, condition = NULL, padj = FALSE)
}
\arguments{
\item{object}{a StatomaticDataset}

\item{condition}{a list which maps columns of get_counts(sds) to experimental factors. If not specified, defaults to NULL and will use the design
formula of sds}

\item{padj}{boolean, whether or not to include Benjamini Hochberg correction. defaults to FALSE}
}
\value{
a description of the method
}
\description{
Runs wilcox test on data contained in a StatomaticDataset
}
