% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapiro_wilk.R
\name{shapiro_wilk}
\alias{shapiro_wilk}
\title{shapiro_wilk}
\usage{
shapiro_wilk(data)
}
\arguments{
\item{data, }{DATAFRAME where data for each continous variable is in its respective column}
}
\value{
LISTS where the first list contains the test statistics and the second list contains the p-values in the order of the continuous variables in the dataframe
}
\description{
Conduct the Shapiro-Wilk test for every continuous variable in the data to test against normality
}
\examples{
iris_data <- data.frame("length" = c(1,2,3,4), "width" = c(5,6,7,8))
shapiro_wilk(iris_data)

}
