% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normtest.R, R/normtest_2.R
\name{summary.normtest}
\alias{summary.normtest}
\title{Summary Method for Objects of Class normtest}
\usage{
\method{summary}{normtest}(x)

\method{summary}{normtest}(x)
}
\arguments{
\item{x}{Object of Class normtest}

\item{x}{Object of Class normtest}
}
\value{
invisible NULL

invisible NULL
}
\description{
Summary Method for Objects of Class normtest

Summary Method for Objects of Class normtest
}
