% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds.R
\name{takeN}
\alias{takeN}
\title{Take n elements of a list and concatenate them. Will introduce NAs if length(x) \%\% n != 0}
\usage{
takeN(x, n = 1, sep = "")
}
\arguments{
\item{x}{a list}

\item{n}{an integer specifying the number of elements to concatenate together}

\item{sep}{a string, the separating string between elements once concatenated}
}
\description{
Take n elements of a list and concatenate them. Will introduce NAs if length(x) \%\% n != 0
}
