% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testnorm.R
\name{test_norm}
\alias{test_norm}
\title{Sorts data based on the distribution of it's residuals, as calculated by glm(family = gaussian), using shapiro and levene tests}
\usage{
test_norm(my_data, var1, var2 = NULL)
}
\arguments{
\item{my_data}{a numeric data.frame}

\item{var1}{a list mapping columns of my_data to experimental factors}

\item{var2}{same as var1, defaults to NULL}
}
\value{
a list containing three data.frames, which are subsets of my_data. Access with double brackets.
[\link{1}] is data which has normally distributed residuals and equal variance between groups, [\link{2}] has normally distributed residuals
and unequal variance between groups, [\link{3}] has non-normally distributed residuals.
}
\description{
Sorts data based on the distribution of it's residuals, as calculated by glm(family = gaussian), using shapiro and levene tests
}
