% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds.R
\name{two_group_fc}
\alias{two_group_fc}
\title{Two group fold changes}
\usage{
two_group_fc(my_data, var1)
}
\arguments{
\item{my_data}{a data object}

\item{var1}{a list that points column names of my_data to the factors of the experiment}
}
\value{
a dataframe of foldchanges
}
\description{
Two group fold changes
}
