% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_group_rui.R
\name{two_group_rui}
\alias{two_group_rui}
\title{This function puts together all the two group functions from the package in the correct order.}
\usage{
two_group_rui(my_data, var1)
}
\arguments{
\item{my_data}{a data object}

\item{var1}{a list that points column names of my_data to the factors of the experiment}
}
\description{
This function puts together all the two group functions from the package in the correct order.
}
