% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_group_test_norm.R
\name{two_group_test_norm}
\alias{two_group_test_norm}
\title{For exactly two groups. Test and sort data by distribution and equality of variances}
\usage{
two_group_test_norm(my_data, var1)
}
\arguments{
\item{my_data}{a data object}

\item{var1}{a list that points column names of my_data to the factors of the experiment}
}
\description{
For exactly two groups. Test and sort data by distribution and equality of variances
}
