% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normtest.R
\name{ad}
\alias{ad}
\title{Anderson and Friend test}
\usage{
ad(x)
}
\arguments{
\item{x}{Numeric Vector}
}
\value{
list
}
\description{
Anderson and Friend test
}
\examples{
set.seed(666)
x <- rnorm(10, 7, 1)
ad(x)
}
