% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normtest.R
\name{basic_stats}
\alias{basic_stats}
\title{Basic Statistics}
\usage{
basic_stats(x, ...)
}
\arguments{
\item{x}{Numeric Vector}

\item{...}{Any parameters to be passed in a function}
}
\value{
Numeric Vector of length six: mean, sd, skewness, kurtosis, n, n_na
}
\description{
Basic Statistics
}
\examples{
set.seed(666)
x <- rnorm(10, 7, 1)
basic_stats(x)
}
