% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds.R
\name{fold_change}
\alias{fold_change}
\title{The fold_change functoin iterates through the groups specified by var1
to calculate the log2FoldChange of mean(groupA) / mean(groupB)}
\usage{
fold_change(my_data, var1)
}
\arguments{
\item{my_data}{a dataframe, double, list with numeric entries}

\item{var1}{a list that points column names of my_data to the factors of the experiment}
}
\value{
a dataframe containing foldchanges for each groupwise comparison
}
\description{
The fold_change functoin iterates through the groups specified by var1
to calculate the log2FoldChange of mean(groupA) / mean(groupB)
}
