% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normtest.R
\name{ks}
\alias{ks}
\title{Kolmogorov-Smirnov test}
\usage{
ks(x, ...)
}
\arguments{
\item{x}{Numeric Vector}

\item{...}{Any extra parameters}
}
\value{
list
}
\description{
Kolmogorov-Smirnov test
}
\examples{
set.seed(666)
x <- rnorm(10, 7, 1)
ks(x)
}
