% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normtest.R
\name{normtest}
\alias{normtest}
\title{Main Normtest Function}
\usage{
normtest(x, ...)
}
\arguments{
\item{x}{Numeric Vector}

\item{...}{Any argument for inner functions basic_stats and tests}
}
\value{
List of class normtest
}
\description{
Main Normtest Function
}
\examples{
# basic examples
set.seed(101)
x <- rnorm(10^2, 7, 1)
test <- normtest(x)
test
summary(test)
# example with NA
y <- x; y[7] <- NA
normtest(y, na.rm = TRUE)
# not normal data
z <- rlnorm(100, 7, 2)
normtest(z)
}
