% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normtest.R
\name{print.normtest}
\alias{print.normtest}
\title{Print method for object of class normtest}
\usage{
\method{print}{normtest}(x)
}
\arguments{
\item{x}{Object of class normtest}
}
\value{
invisible NULL
}
\description{
Print method for object of class normtest
}
