% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_tukey.R
\name{run_anova}
\alias{run_anova}
\title{Runs a two way anova based on var1 * var2 for each row of my_data}
\usage{
run_anova(my_data, var1, var2)
}
\arguments{
\item{my_data}{a data object}

\item{var1}{a list that points column names of my_data to the factors of the experiment}

\item{var2}{a list that points column names of my_data to the factors of the experiment}
}
\value{
a data object containing p values for var1, var2, and interaction of var1 * var2 for each row of my_data, computed by anova
}
\description{
Runs a two way anova based on var1 * var2 for each row of my_data
}
