% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kruskalwallis_dunn.R
\name{run_dunn}
\alias{run_dunn}
\title{Runs Dunn's multi comparison test on each row of my_data}
\usage{
run_dunn(my_data, var1)
}
\arguments{
\item{my_data}{a data object}

\item{var1}{a list that points column names of my_data to the factors of the experiment}
}
\value{
a data object with a pvalue for each groupwise comparison, as specified by var1, computed using an altered version of dunn.test
}
\description{
Runs Dunn's multi comparison test on each row of my_data
}
