% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kruskalwallis_dunn.R
\name{run_kruskal}
\alias{run_kruskal}
\title{Run Kruskal-Wallis test on each row of my_data}
\usage{
run_kruskal(my_data, var1)
}
\arguments{
\item{my_data}{a data object}

\item{var1}{a list that points column names of my_data to the factors of the experiment}
}
\value{
a data object with a p value for each row of my_data, computed using kruskal.test
}
\description{
Run Kruskal-Wallis test on each row of my_data
}
