% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ttest.R
\name{run_ttest}
\alias{run_ttest}
\title{Runs an unpaired t-test on each row of my_data and reports the p value, computed by t.test}
\usage{
run_ttest(my_data, var1)
}
\arguments{
\item{my_data}{a data object}

\item{var1}{list that points column names of my_data to the factors of the experiment}
}
\description{
Runs an unpaired t-test on each row of my_data and reports the p value, computed by t.test
}
