% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tukey_test.R
\name{run_tukey}
\alias{run_tukey}
\title{Performs Tukey's HSD test on each row of a data.frame}
\usage{
run_tukey(my_data, var1)
}
\arguments{
\item{my_data}{a numeric data.frame}

\item{var1}{a list which maps columns of my_data to experimental factors}
}
\value{
a data frame of p values
}
\description{
Performs Tukey's HSD test on each row of a data.frame
}
