% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/welch_test.R
\name{run_welch}
\alias{run_welch}
\title{Performs Welch t-test on each row of a data.frame}
\usage{
run_welch(my_data, var1, padj = FALSE, include_test_name = FALSE)
}
\arguments{
\item{my_data}{a numeric data.frame}

\item{var1}{a list which maps columns of my_data to experimental factors}

\item{padj}{boolean, whether or not to include Benjamini Hochberg correction. defaults to FALSE}
}
\value{
a data.frame of p values
}
\description{
Performs Welch t-test on each row of a data.frame
}
