% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normtest.R
\name{summary.normtest}
\alias{summary.normtest}
\title{Summary method for object of class normtest}
\usage{
\method{summary}{normtest}(x)
}
\arguments{
\item{x}{Object of class normtest}
}
\value{
invisible NULL
}
\description{
Summary method for object of class normtest
}
