% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_norm.R
\name{test_norm}
\alias{test_norm}
\title{For at least 3 groups. Test and sort data by distribution and variance equality}
\usage{
test_norm(my_data, var1, var2)
}
\arguments{
\item{my_data}{A numeric set of data such as list, double, integer, etc.}

\item{var1}{A column from a sample_info table used to define which column names of my_data go with var1}

\item{var2}{A column from a sample_info table used to define which column names of my_data go with var2}
}
\value{
Three dataframes ne, nu, nn assigned to global environment
}
\description{
For at least 3 groups. Test and sort data by distribution and variance equality
}
