% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twogroup_main.R
\name{twogroup_main}
\alias{twogroup_main}
\title{This function puts together all the two group functions from the package in the correct order.}
\usage{
twogroup_main(my_data, var1, padj = FALSE)
}
\arguments{
\item{my_data}{a data object}

\item{var1}{a list that points column names of my_data to the factors of the experiment}
}
\description{
This function puts together all the two group functions from the package in the correct order.
}
